merge origin/master into feat/1237 (resolves loraadapterbase conflict
by taking master's non-virtual `packbaseandloraparameters` refactor and
casting `parametercount` to int in `rebuildparametersafterderivedinit`
since `vector.length` stays int per the per-tensor contract).

audit gaps closed:

1. **expectedparametercount widened to long**. classifierbase and
   clusteringbase compute `numfeatures * numclasses` /
   `numfeatures * numclusters` and surface it through
   `parametercount`. on huge classifier configs (large vocab classifiers,
   high-dim feature spaces) the int product overflows even though
   parametercount itself is long. cast operands to long for the
   multiplication.

2. **internal int accumulation widened to long in foundation-scale
   predictors**. `ditnoisepredictor.parametercount` and
   `mmditnoisepredictor.calculateparametercount` were summing
   per-layer counts into `int count = 0` then implicitly converting to
   long on return — silent overflow at sora (5.4 b) and sd3.5 large
   (7.6 b) scales. switch to `long count = 0` and drop the redundant
   `(int)` casts on per-layer parametercount reads (they're long now,
   so no explicit conversion needed for the running sum).

3. **composite gan getparameterchunks overrides** added per the issue's
   scope #3 ("composite networks (cyclegan, infogan, gan sub-network
   containers) → forward to sub-networks"):
   - `cyclegan.getparameterchunks()` — yields from generatoratob,
     generatorbtoa, discriminatora, discriminatorb in turn
   - `infogan.getparameterchunks()` — yields from generator, discriminator,
     qnetwork in turn
   - `generativeadversarialnetwork.getparameterchunks()` — yields from
     generator then discriminator (covers biggan, progressivegan, sagan
     since they extend gan and inherit this override)

build status:
- src/aidotnet.csproj: clean (0 errors)
- tests/aidotnet.tests.csproj: clean (0 errors)
- soramodel_haspaperfaithfulcomponents: passes in 12 s

+++ src/Classification/ClassifierBase.cs
@@ -130,7 +130,7 @@ public abstract class ClassifierBase<T> : IClassifier<T>, IConfigurableModel<T>,
     /// <value>
     /// The total number of parameters in the model, used for serialization and gradient computation.
     /// </value>
-    protected virtual int ExpectedParameterCount => NumFeatures * NumClasses;
+    protected virtual long ExpectedParameterCount => (long)NumFeatures * NumClasses;
 
     /// <summary>
     /// Initializes a new instance of the ClassifierBase class with the specified options and regularization.
